+++ ios/PinayPalBackup/PinayPalBackupApp.swift
@@ -5,6 +5,10 @@ struct PinayPalBackupApp: App {
     var body: some Scene {
         WindowGroup {
             MainView()
+                .task {
+                    _ = await NotificationService.shared.requestAuthorization()
+                    BackupLiveActivityManager.shared.restoreActiveActivityIfNeeded()
+                }
         }
     }
 }